+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Added Default `DateFormat` as `YYYY-MM-DD` in the `AppDatePicker`.
 - Added empty string for `labelText` in the `AppDatePicker`.
 - Added `MultiEnumOptionCubit` for managing multi-select enum fields.
+- Added `EnumMultiOptionCheckboxGroup` a Multi-select checkbox group for List of Enums.
 
 ## 1.2.9